dofunc/packages/bhrtools/bhrteamcal/requirements.txt
requests
icalendar
boto3
python-dotenv
pyyaml